+++ NEWS.md
@@ -1,6 +1,7 @@
 # Unreleased version
 
 - Fix bug: `disable()` now works with file inputs (#249)
+- Added support for `reset()` function to work with dates and date-times in slider inputs (#213)
 - Added `shinyjs.version` JavaScript variable to help troubleshooting
 - Replace `digest::digest()` with `rlang::hash`, to reduce dependencies and it's faster (#248)
 - Replace `list(...)` with `rlang::list2(...)` which is more flexible and allows you to use trailing commas without error